Default covariates for use in augury functions.

A dataset containing default covariates for all WHO member states, for use within augury functions. These are used if no defaults are provided to the modeling functions.

Usage

covariates_df

Format

A data frame with 5044 rows and 11 variables:

iso3

Country ISO3 code.

year

Year.

year_n

Year, starting with 1 in 2000. For use in time series modelling.

region

Unclear origin regional classification.

un_subregion

UN subregional classification.

sdg_subregion

UN subregional classification for SDG calculation purposes.

gbd_subregion

IHME GBD subregional classification.

sdi

Socio-demographic Index, a composite calculated using publicly available on total fertility rates, mean years of schooling, and lag distributed income per capita.

sdi_scaled

Socio-demographic Index, rescaled to a 0 - 1 scale.

e0

Life expectancy at birth, total (years).

e0_scaled

Life expectancy at birth, rescaled to a 0 - 1 scale.

Source

Region, World Bank regional groups, 2020.

Socio-demographic Index, World Health Organization calculation, 2020.

Life expectancy, World Population Prospects, 2019.